working on it
extended content on both blog and merch pages.

fiddled with footer placement

fiddled with general spacing.

added some experimental animation for the product page. it seems that i would need to use javascript to make a smooth transition back to the default state, so i filled in a second animation...? it somewhat conceals the choppy transition, definitely isn't perfect or ideal, but it's fun to play with and run your mouse over.

related to above: provided general hover effects for products. added minor transition animation because i'm obsessed with the effect.