created template layouts
created template layouts for log and calendar. this involved setting up css for the layouts themselves and for card displays [and buttons.

i choose to do this first because once i know what i want my javascript to build, it will be easier to verify if i've written my js right or not. additionally, it gets a big piece of the puzzle out of the way in my mind.

the next step is to build the entry creation page and its functionality. i can use it to create journal entries, and use the console to verify things are structured correctly. then i can work back towards building the layout pages with the given data.